remove special defaultvalidateexecutor type and trait
